Updated Nav, fonts, and colors
There was an issue that held me back for a while when I was working on the fonts. I incorrectly had a period before my 'truetype' and 'woff' formats. Next time I work on this I need to add more comments. Otherwise, this focused on the looks of the first project.